+++ controllers/s3/factory/minioS3Client.go
@@ -21,13 +21,13 @@ func (minioS3Client *MinioS3Client) BucketExists(name string) (bool, error) {
 	return minioS3Client.client.BucketExists(context.Background(), name)
 }
 
-func (minioS3Client *MinioS3Client) GetQuota(name string) (int32, error) {
+func (minioS3Client *MinioS3Client) GetQuota(name string) (int64, error) {
 	log.Println("bucket " + name + " get quota")
 	bucketQuota, err := minioS3Client.adminClient.GetBucketQuota(context.Background(), name)
 	if err != nil {
 		log.Fatalln(err)
 	}
-	return int32(bucketQuota.Quota), err
+	return int64(bucketQuota.Quota), err
 }
 
 func (minioS3Client *MinioS3Client) CreateBucket(name string) error {
@@ -40,7 +40,7 @@ func (minioS3Client *MinioS3Client) DeleteBucket(name string) error {
 	return minioS3Client.client.RemoveBucket(context.Background(), name)
 }
 
-func (minioS3Client *MinioS3Client) SetQuota(name string, quota int32) error {
+func (minioS3Client *MinioS3Client) SetQuota(name string, quota int64) error {
 	log.Println("set quota " + fmt.Sprint(quota) + "on bucket " + name + "exists")
 	minioS3Client.adminClient.SetBucketQuota(context.Background(), name, &madmin.BucketQuota{Quota: uint64(quota), Type: madmin.HardQuota})
 	return nil
